Refactor: API 모듈 AuctionService 수정 - 가격 타입 Long, Int 등 불일치했던 것들을 Int 타입으로 수정 - AuctionServiceTest 코드 보완 - 경매 최초입찰 처리 보완

+++ module-api/src/main/java/com/example/moduleapi/service/auction/KafkaProducerService.java
@@ -14,11 +14,11 @@ public class KafkaProducerService {
     }
 
     @Transactional
-    public void publishAuctionPriceChangeNotification(Long productId, Long newPrice) {
+    public void publishAuctionPriceChangeNotification(Long productId, int newPrice) {
         AuctionPriceChangeNotification notification = AuctionPriceChangeNotification.builder()
-                .productId(productId)
-                .newPrice(newPrice)
-                .build();
+                                                                                    .productId(productId)
+                                                                                    .newPrice(newPrice)
+                                                                                    .build();
 
         kafkaTemplate.send("Auction-Price-Change-Notification", notification);
